refactor: 사물함 feature 경로 정리 #53
